refactor: rename project-data-service.error to project-error
