+++ server/src/features/projects/services/project-data-service/index.ts
@@ -1,3 +1,3 @@
-export * from "./project-data-service.error";
+export * from "./project-error";
 export * from "./project-data-service";
 export * from "./project-data-service.type";
